Assign to me on new issues

+++ internal/cmd/issue/structs.go
@@ -30,7 +30,7 @@ type newIssue struct {
 	Description    string `json:"description,omitempty"`
 	CategoryID     int64  `json:"category_id,omitempty"`
 	FixedVersionID int64  `json:"fixed_version_id,omitempty"`
-	AssignedToID   int64  `json:"assigned_to_id,omitempty"`
+	AssignedToID   int    `json:"assigned_to_id,omitempty"`
 	ParentIssueID  int64  `json:"parent_issue_id,omitempty"`
 	Private        bool   `json:"is_private,omitempty"`
 	EstimatedHours int    `json:"estimated_hours,omitempty"`
